agent/requirements.txt
fastapi
uvicorn
langchain
langchain-anthropic
requests
python-dotenv
